Fix tests for tsdx
