+++ jest.config.js
@@ -1,8 +1,4 @@
 module.exports = {
-  collectCoverage: true,
-  collectCoverageFrom: ['src/**/*.{ts,tsx}', '!**/node_modules/**'],
-  coverageDirectory: 'coverage',
-  testURL: 'http://localhost/',
-  testEnvironment: 'jsdom',
   setupFilesAfterEnv: ['./jest-setup.js'],
+  testMatch: ['<rootDir>/src/**/*.(spec|test).{ts,tsx,js,jsx}'],
 };
