+++ src/focus-node.test.tsx
@@ -1,3 +1,4 @@
+// @ts-nocheck
 import React, { useRef } from 'react';
 import '@testing-library/jest-dom';
 import { render, act, fireEvent, screen } from '@testing-library/react';
