+++ src/hooks/use-focus-store-dangerously.test.tsx
@@ -1,3 +1,4 @@
+// @ts-nocheck
 import React from 'react';
 import '@testing-library/jest-dom';
 import { render, act } from '@testing-library/react';
