+++ src/tests/pointer-events.test.tsx
@@ -1,3 +1,4 @@
+// @ts-nocheck
 import React from 'react';
 import '@testing-library/jest-dom';
 import { render, fireEvent, createEvent, screen } from '@testing-library/react';
