+++ src/utils/warning.test.ts
@@ -1,3 +1,4 @@
+// @ts-nocheck
 import { warning } from './warning';
 
 describe('warning', () => {
