+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
   // see https://www.typescriptlang.org/tsconfig to better understand tsconfigs
-  "include": ["src", "types"],
+  "include": ["src", "types", "./jest-setup.js"],
   "compilerOptions": {
     "module": "esnext",
     "lib": ["dom", "esnext"],
